requirements.txt
python-telegram-bot==21.4
spacy==3.7.5
spacy==3.7.6
